testapp/requirements.txt
django
djangorestframework
coverage
flake8
tox
